+++ src/client/messages/oldcommands/InternCommand.java
@@ -433,15 +433,6 @@ public class InternCommand {
         }
     }
 
-    public static class SpawnDebug extends CommandExecute {
-
-        @Override
-        public int execute(MapleClient c, String[] splitted) {
-            c.getPlayer().dropMessage(6, c.getPlayer().getMap().spawnDebug());
-            return 1;
-        }
-    }
-
     public static class FakeRelog extends CommandExecute {
 
         @Override
@@ -488,14 +479,6 @@ public class InternCommand {
         }
     }
 
-    public static class Uptime extends CommandExecute {
-
-        @Override
-        public int execute(MapleClient c, String[] splitted) {
-            c.getPlayer().dropMessage(6, "Server has been up for " + StringUtil.getReadableMillis(ChannelServer.serverStartTime, System.currentTimeMillis()));
-            return 1;
-        }
-    }
 
     public static class EventInstance extends CommandExecute {
 
@@ -511,214 +494,11 @@ public class InternCommand {
         }
     }
 
-    public static class GoTo extends CommandExecute {
-
-        private static final HashMap<String, Integer> gotomaps = new HashMap<String, Integer>();
-
-        static {
-            gotomaps.put("gmmap", 180000000);
-            gotomaps.put("southperry", 2000000);
-            gotomaps.put("amherst", 1010000);
-            gotomaps.put("henesys", 100000000);
-            gotomaps.put("ellinia", 101000000);
-            gotomaps.put("perion", 102000000);
-            gotomaps.put("kerning", 103000000);
-            gotomaps.put("harbor", 104000000);
-            gotomaps.put("sleepywood", 105000000);
-            gotomaps.put("florina", 120000300);
-            gotomaps.put("orbis", 200000000);
-            gotomaps.put("happyville", 209000000);
-            gotomaps.put("elnath", 211000000);
-            gotomaps.put("ludibrium", 220000000);
-            gotomaps.put("aquaroad", 230000000);
-            gotomaps.put("leafre", 240000000);
-            gotomaps.put("mulung", 250000000);
-            gotomaps.put("herbtown", 251000000);
-            gotomaps.put("omegasector", 221000000);
-            gotomaps.put("koreanfolktown", 222000000);
-            gotomaps.put("newleafcity", 600000000);
-            gotomaps.put("sharenian", 990000000);
-            gotomaps.put("pianus", 230040420);
-            gotomaps.put("horntail", 240060200);
-            gotomaps.put("chorntail", 240060201);
-            gotomaps.put("griffey", 240020101);
-            gotomaps.put("manon", 240020401);
-            gotomaps.put("zakum", 280030000);
-            gotomaps.put("czakum", 280030001);
-            gotomaps.put("papulatus", 220080001);
-            gotomaps.put("showatown", 801000000);
-            gotomaps.put("zipangu", 800000000);
-            gotomaps.put("ariant", 260000100);
-            gotomaps.put("nautilus", 120000000);
-            gotomaps.put("boatquay", 541000000);
-            gotomaps.put("malaysia", 550000000);
-            gotomaps.put("erev", 130000000);
-            gotomaps.put("ellin", 300000000);
-            gotomaps.put("kampung", 551000000);
-            gotomaps.put("singapore", 540000000);
-            gotomaps.put("amoria", 680000000);
-            gotomaps.put("timetemple", 270000000);
-            gotomaps.put("pinkbean", 270050100);
-            gotomaps.put("fm", 910000000);
-            gotomaps.put("freemarket", 910000000);
-            gotomaps.put("oxquiz", 109020001);
-            gotomaps.put("ola", 109030101);
-            gotomaps.put("fitness", 109040000);
-            gotomaps.put("snowball", 109060000);
-            gotomaps.put("golden", 950100000);
-            gotomaps.put("phantom", 610010000);
-            gotomaps.put("cwk", 610030000);
-            gotomaps.put("rien", 140000000);
-            gotomaps.put("edel", 310000000);
-            gotomaps.put("ardent", 910001000);
-            gotomaps.put("craft", 910001000);
-            gotomaps.put("pvp", 960000000);
-            gotomaps.put("future", 271000000);
-        }
-
-        @Override
-        public int execute(MapleClient c, String[] splitted) {
-            if (splitted.length < 2) {
-                c.getPlayer().dropMessage(6, "Syntax: !goto <mapname>");
-            } else {
-                if (gotomaps.containsKey(splitted[1])) {
-                    MapleMap target = c.getChannelServer().getMapFactory().getMap(gotomaps.get(splitted[1]));
-                    if (target == null) {
-                        c.getPlayer().dropMessage(6, "Map does not exist");
-                        return 0;
-                    }
-                    MaplePortal targetPortal = target.getPortal(0);
-                    c.getPlayer().changeMap(target, targetPortal);
-                } else {
-                    if (splitted[1].equals("locations")) {
-                        c.getPlayer().dropMessage(6, "Use !goto <location>. Locations are as follows:");
-                        StringBuilder sb = new StringBuilder();
-                        for (String s : gotomaps.keySet()) {
-                            sb.append(s).append(", ");
-                        }
-                        c.getPlayer().dropMessage(6, sb.substring(0, sb.length() - 2));
-                    } else {
-                        c.getPlayer().dropMessage(6, "Invalid command syntax - Use !goto <location>. For a list of locations, use !goto locations.");
-                    }
-                }
-            }
-            return 1;
-        }
-    }
-
-    public static class MonsterDebug extends CommandExecute {
-
-        @Override
-        public int execute(MapleClient c, String[] splitted) {
-            MapleMap map = c.getPlayer().getMap();
-            double range = Double.POSITIVE_INFINITY;
-
-            if (splitted.length > 1) {
-                //&& !splitted[0].equals("!killmonster") && !splitted[0].equals("!hitmonster") && !splitted[0].equals("!hitmonsterbyoid") && !splitted[0].equals("!killmonsterbyoid")) {
-                int irange = Integer.parseInt(splitted[1]);
-                if (splitted.length <= 2) {
-                    range = irange * irange;
-                } else {
-                    map = c.getChannelServer().getMapFactory().getMap(Integer.parseInt(splitted[2]));
-                }
-            }
-            if (map == null) {
-                c.getPlayer().dropMessage(6, "Map does not exist");
-                return 0;
-            }
-            MapleMonster mob;
-            for (MapleMapObject monstermo : map.getMapObjectsInRange(c.getPlayer().getPosition(), range, Arrays.asList(MapleMapObjectType.MONSTER))) {
-                mob = (MapleMonster) monstermo;
-                c.getPlayer().dropMessage(6, "Monster " + mob.toString());
-            }
-            return 1;
-        }
-    }
-
-    public static class LookNPC extends CommandExecute {
-
-        @Override
-        public int execute(MapleClient c, String[] splitted) {
-            for (MapleMapObject reactor1l : c.getPlayer().getMap().getAllNPCsThreadsafe()) {
-                MapleNPC reactor2l = (MapleNPC) reactor1l;
-                c.getPlayer().dropMessage(5, "NPC: oID: " + reactor2l.getObjectId() + " npcID: " + reactor2l.getId() + " Position: " + reactor2l.getPosition().toString() + " Name: " + reactor2l.getName());
-            }
-            return 0;
-        }
-    }
-
-    public static class LookReactor extends CommandExecute {
-
-        @Override
-        public int execute(MapleClient c, String[] splitted) {
-            for (MapleMapObject reactor1l : c.getPlayer().getMap().getAllReactorsThreadsafe()) {
-                MapleReactor reactor2l = (MapleReactor) reactor1l;
-                c.getPlayer().dropMessage(5, "Reactor: oID: " + reactor2l.getObjectId() + " reactorID: " + reactor2l.getReactorId() + " Position: " + reactor2l.getPosition().toString() + " State: " + reactor2l.getState() + " Name: " + reactor2l.getName());
-            }
-            return 0;
-        }
-    }
-
-    public static class LookPortals extends CommandExecute {
-
-        @Override
-        public int execute(MapleClient c, String[] splitted) {
-            for (MaplePortal portal : c.getPlayer().getMap().getPortals()) {
-                c.getPlayer().dropMessage(5, "Portal: ID: " + portal.getId() + " script: " + portal.getScriptName() + " name: " + portal.getName() + " pos: " + portal.getPosition().x + "," + portal.getPosition().y + " target: " + portal.getTargetMapId() + " / " + portal.getTarget());
-            }
-            return 0;
-        }
-    }
-
-    public static class MyNPCPos extends CommandExecute {
-
-        @Override
-        public int execute(MapleClient c, String[] splitted) {
-            Point pos = c.getPlayer().getPosition();
-            c.getPlayer().dropMessage(6, "X: " + pos.x + " | Y: " + pos.y + " | RX0: " + (pos.x + 50) + " | RX1: " + (pos.x - 50) + " | FH: " + c.getPlayer().getFH());
-            return 1;
-        }
-    }
-
     public static class Clock extends CommandExecute {
 
         @Override
         public int execute(MapleClient c, String[] splitted) {
             c.getPlayer().getMap().broadcastMessage(CField.getClock(CommandProcessorUtil.getOptionalIntArg(splitted, 1, 60)));
-            return 1;
-        }
-    }
-
-    public static class WarpHere extends CommandExecute {
-
-        @Override
-        public int execute(MapleClient c, String[] splitted) {
-            MapleCharacter victim = c.getChannelServer().getPlayerStorage().getCharacterByName(splitted[1]);
-            if (victim != null) {
-                if (c.getPlayer().inPVP() || (!c.getPlayer().isGM() && (victim.isInBlockedMap() || victim.isGM()))) {
-                    c.getPlayer().dropMessage(5, "Try again later.");
-                    return 0;
-                }
-                victim.changeMap(c.getPlayer().getMap(), c.getPlayer().getMap().findClosestPortal(c.getPlayer().getTruePosition()));
-            } else {
-                int ch = World.Find.findChannel(splitted[1]);
-                if (ch < 0) {
-                    c.getPlayer().dropMessage(5, "Not found.");
-                    return 0;
-                }
-                victim = ChannelServer.getInstance(ch).getPlayerStorage().getCharacterByName(splitted[1]);
-                if (victim == null || victim.inPVP() || (!c.getPlayer().isGM() && (victim.isInBlockedMap() || victim.isGM()))) {
-                    c.getPlayer().dropMessage(5, "Try again later.");
-                    return 0;
-                }
-                c.getPlayer().dropMessage(5, "Victim is cross changing channel.");
-                victim.dropMessage(5, "Cross changing channel.");
-                if (victim.getMapId() != c.getPlayer().getMapId()) {
-                    final MapleMap mapp = victim.getClient().getChannelServer().getMapFactory().getMap(c.getPlayer().getMapId());
-                    victim.changeMap(mapp, mapp.findClosestPortal(c.getPlayer().getTruePosition()));
-                }
-                victim.changeChannel(c.getChannel());
-            }
             return 1;
         }
     }
@@ -796,63 +576,6 @@ public class InternCommand {
         }
     }
 
-    public static class Jail extends CommandExecute {
-
-        @Override
-        public int execute(MapleClient c, String[] splitted) {
-            if (splitted.length < 3) {
-                c.getPlayer().dropMessage(6, "jail [name] [minutes, 0 = forever]");
-                return 0;
-            }
-            MapleCharacter victim = c.getChannelServer().getPlayerStorage().getCharacterByName(splitted[1]);
-            final int minutes = Math.max(0, Integer.parseInt(splitted[2]));
-            if (victim != null && c.getPlayer().getGMLevel() >= victim.getGMLevel()) {
-                MapleMap target = ChannelServer.getInstance(c.getChannel()).getMapFactory().getMap(GameConstants.JAIL);
-                victim.getQuestNAdd(MapleQuest.getInstance(GameConstants.JAIL_QUEST)).setCustomData(String.valueOf(minutes * 60));
-                victim.changeMap(target, target.getPortal(0));
-            } else {
-                c.getPlayer().dropMessage(6, "Please be on their channel.");
-                return 0;
-            }
-            return 1;
-        }
-    }
-
-    public static class ListAllSquads extends CommandExecute {
-
-        @Override
-        public int execute(MapleClient c, String[] splitted) {
-            for (ChannelServer cserv : ChannelServer.getAllInstances()) {
-                for (Entry<MapleSquad.MapleSquadType, MapleSquad> squads : cserv.getAllSquads().entrySet()) {
-                    c.getPlayer().dropMessage(5, "[Channel " + cserv.getChannel() + "] TYPE: " + squads.getKey().name() + ", Leader: " + squads.getValue().getLeader().getName() + ", status: " + squads.getValue().getStatus() + ", numMembers: " + squads.getValue().getSquadSize() + ", numBanned: " + squads.getValue().getBannedMemberSize());
-                }
-            }
-            return 1;
-        }
-    }
-
-    public static class Say extends CommandExecute {
-
-        @Override
-        public int execute(MapleClient c, String[] splitted) {
-            if (splitted.length > 1) {
-                StringBuilder sb = new StringBuilder();
-                sb.append("[");
-                if (!c.getPlayer().isGM()) {
-                    sb.append("Intern ");
-                }
-                sb.append(c.getPlayer().getName());
-                sb.append("] ");
-                sb.append(StringUtil.joinStringFrom(splitted, 1));
-                World.Broadcast.broadcastMessage(CWvsContext.serverNotice(c.getPlayer().isGM() ? 6 : 5, sb.toString()));
-            } else {
-                c.getPlayer().dropMessage(6, "Syntax: say <message>");
-                return 0;
-            }
-            return 1;
-        }
-    }
-
     public static class Letter extends CommandExecute {
 
         @Override
@@ -905,141 +628,6 @@ public class InternCommand {
             }
             return 1;
         }
-    }
-
-    public static class Find extends CommandExecute {
-
-        @Override
-        public int execute(MapleClient c, String[] splitted) {
-            if (splitted.length == 1) {
-                c.getPlayer().dropMessage(6, splitted[0] + ": <NPC> <MOB> <ITEM> <MAP> <SKILL> <QUEST>");
-            } else if (splitted.length == 2) {
-                c.getPlayer().dropMessage(6, "Provide something to search.");
-            } else {
-                String type = splitted[1];
-                String search = StringUtil.joinStringFrom(splitted, 2);
-                MapleData data = null;
-                MapleDataProvider dataProvider = MapleDataProviderFactory.getDataProvider(new File(System.getProperty("net.sf.odinms.wzpath") + "/" + "String.wz"));
-                c.getPlayer().dropMessage(6, "<<Type: " + type + " | Search: " + search + ">>");
-
-                if (type.equalsIgnoreCase("NPC")) {
-                    List<String> retNpcs = new ArrayList<String>();
-                    data = dataProvider.getData("Npc.img");
-                    List<Pair<Integer, String>> npcPairList = new LinkedList<Pair<Integer, String>>();
-                    for (MapleData npcIdData : data.getChildren()) {
-                        npcPairList.add(new Pair<Integer, String>(Integer.parseInt(npcIdData.getName()), MapleDataTool.getString(npcIdData.getChildByPath("name"), "NO-NAME")));
-                    }
-                    for (Pair<Integer, String> npcPair : npcPairList) {
-                        if (npcPair.getRight().toLowerCase().contains(search.toLowerCase())) {
-                            retNpcs.add(npcPair.getLeft() + " - " + npcPair.getRight());
-                        }
-                    }
-                    if (retNpcs != null && retNpcs.size() > 0) {
-                        for (String singleRetNpc : retNpcs) {
-                            c.getPlayer().dropMessage(6, singleRetNpc);
-                        }
-                    } else {
-                        c.getPlayer().dropMessage(6, "No NPC's Found");
-                    }
-
-                } else if (type.equalsIgnoreCase("MAP")) {
-                    List<String> retMaps = new ArrayList<String>();
-                    data = dataProvider.getData("Map.img");
-                    List<Pair<Integer, String>> mapPairList = new LinkedList<Pair<Integer, String>>();
-                    for (MapleData mapAreaData : data.getChildren()) {
-                        for (MapleData mapIdData : mapAreaData.getChildren()) {
-                            mapPairList.add(new Pair<Integer, String>(Integer.parseInt(mapIdData.getName()), MapleDataTool.getString(mapIdData.getChildByPath("streetName"), "NO-NAME") + " - " + MapleDataTool.getString(mapIdData.getChildByPath("mapName"), "NO-NAME")));
-                        }
-                    }
-                    for (Pair<Integer, String> mapPair : mapPairList) {
-                        if (mapPair.getRight().toLowerCase().contains(search.toLowerCase())) {
-                            retMaps.add(mapPair.getLeft() + " - " + mapPair.getRight());
-                        }
-                    }
-                    if (retMaps != null && retMaps.size() > 0) {
-                        for (String singleRetMap : retMaps) {
-                            c.getPlayer().dropMessage(6, singleRetMap);
-                        }
-                    } else {
-                        c.getPlayer().dropMessage(6, "No Maps Found");
-                    }
-                } else if (type.equalsIgnoreCase("MOB")) {
-                    List<String> retMobs = new ArrayList<String>();
-                    data = dataProvider.getData("Mob.img");
-                    List<Pair<Integer, String>> mobPairList = new LinkedList<Pair<Integer, String>>();
-                    for (MapleData mobIdData : data.getChildren()) {
-                        mobPairList.add(new Pair<Integer, String>(Integer.parseInt(mobIdData.getName()), MapleDataTool.getString(mobIdData.getChildByPath("name"), "NO-NAME")));
-                    }
-                    for (Pair<Integer, String> mobPair : mobPairList) {
-                        if (mobPair.getRight().toLowerCase().contains(search.toLowerCase())) {
-                            retMobs.add(mobPair.getLeft() + " - " + mobPair.getRight());
-                        }
-                    }
-                    if (retMobs != null && retMobs.size() > 0) {
-                        for (String singleRetMob : retMobs) {
-                            c.getPlayer().dropMessage(6, singleRetMob);
-                        }
-                    } else {
-                        c.getPlayer().dropMessage(6, "No Mobs Found");
-                    }
-
-                } else if (type.equalsIgnoreCase("ITEM")) {
-                    List<String> retItems = new ArrayList<String>();
-                    for (ItemInformation itemPair : MapleItemInformationProvider.getInstance().getAllItems()) {
-                        if (itemPair != null && itemPair.name != null && itemPair.name.toLowerCase().contains(search.toLowerCase())) {
-                            retItems.add(itemPair.itemId + " - " + itemPair.name);
-                        }
-                    }
-                    if (retItems != null && retItems.size() > 0) {
-                        for (String singleRetItem : retItems) {
-                            c.getPlayer().dropMessage(6, singleRetItem);
-                        }
-                    } else {
-                        c.getPlayer().dropMessage(6, "No Items Found");
-                    }
-                } else if (type.equalsIgnoreCase("QUEST")) {
-                    List<String> retItems = new ArrayList<String>();
-                    for (MapleQuest itemPair : MapleQuest.getAllInstances()) {
-                        if (itemPair.getName().length() > 0 && itemPair.getName().toLowerCase().contains(search.toLowerCase())) {
-                            retItems.add(itemPair.getId() + " - " + itemPair.getName());
-                        }
-                    }
-                    if (retItems != null && retItems.size() > 0) {
-                        for (String singleRetItem : retItems) {
-                            c.getPlayer().dropMessage(6, singleRetItem);
-                        }
-                    } else {
-                        c.getPlayer().dropMessage(6, "No Quests Found");
-                    }
-                } else if (type.equalsIgnoreCase("SKILL")) {
-                    List<String> retSkills = new ArrayList<String>();
-                    for (Skill skil : SkillFactory.getAllSkills()) {
-                        if (skil.getName() != null && skil.getName().toLowerCase().contains(search.toLowerCase())) {
-                            retSkills.add(skil.getId() + " - " + skil.getName());
-                        }
-                    }
-                    if (retSkills != null && retSkills.size() > 0) {
-                        for (String singleRetSkill : retSkills) {
-                            c.getPlayer().dropMessage(6, singleRetSkill);
-                        }
-                    } else {
-                        c.getPlayer().dropMessage(6, "No Skills Found");
-                    }
-                } else {
-                    c.getPlayer().dropMessage(6, "Sorry, that search call is unavailable");
-                }
-            }
-            return 0;
-        }
-    }
-
-    public static class ID extends Find {
-    }
-
-    public static class LookUp extends Find {
-    }
-
-    public static class Search extends Find {
     }
 
     public static class WhosFirst extends CommandExecute {
@@ -1171,33 +759,5 @@ public class InternCommand {
         }
     }
 
-    public static class KillAll extends CommandExecute {
-
-        @Override
-        public int execute(MapleClient c, String[] splitted) {
-            MapleMap map = c.getPlayer().getMap();
-            double range = Double.POSITIVE_INFINITY;
-
-            if (splitted.length > 1) {
-                int irange = Integer.parseInt(splitted[1]);
-                if (splitted.length <= 2) {
-                    range = irange * irange;
-                } else {
-                    map = c.getChannelServer().getMapFactory().getMap(Integer.parseInt(splitted[2]));
-                }
-            }
-            if (map == null) {
-                c.getPlayer().dropMessage(6, "Map does not exist");
-                return 0;
-            }
-            MapleMonster mob;
-            for (MapleMapObject monstermo : map.getMapObjectsInRange(c.getPlayer().getPosition(), range, Arrays.asList(MapleMapObjectType.MONSTER))) {
-                mob = (MapleMonster) monstermo;
-                if (!mob.getStats().isBoss() || mob.getStats().isPartyBonus() || c.getPlayer().isGM()) {
-                    map.killMonster(mob, c.getPlayer(), false, false, (byte) 1);
-                }
-            }
-            return 1;
-        }
-    }
+   
 }*/
